Worked on displaying the options when a option is selected
