+++ scripts/main.js
@@ -1,9 +1,10 @@
 import { buildFacilityDropdown } from "./facilities.js";
 import { governorsHTML } from "./governors.js";
 import {governorChoiceChange} from "./colonyMinerals.js"
+import { showAvailableMinerals } from "./availableMinerals.js";
 
 document.addEventListener("change", governorChoiceChange)
-
+document.addEventListener("change", showAvailableMinerals)
 const createPage = () => {
   const documentHTML = document.querySelector(`#ContainerJS`);
   const pageHtml = `
@@ -13,6 +14,7 @@ const createPage = () => {
             ${governorsHTML()}
             <article id="colonyMinerals"></article>
             ${buildFacilityDropdown()}
+            <article id='facilityMineralsDocument'></article>
             <button id="purchase">Purchase Minerals!</button>
         </div>
     </div>
